Add Logger Feature (Phase1)

+++ TestShell/TestShell/cmd_interface.cpp
@@ -27,7 +27,7 @@ void CmdInterface::checkDataArg(const string& dataString) const {
     if (isValidDataStringLen(dataString))
     {
         std::size_t errorPos = 0;
-        unsigned int data = std::stoi(dataString, &errorPos, 16);
+        unsigned int data = std::stoul(dataString, &errorPos, 16);
         if (isValidDataString(dataString, errorPos))
         {
             if (isValidDataRange(data)) return;
